Rename soap-scum.rb to soap_scum.rb
