+++ lib/vbms.rb
@@ -13,7 +13,7 @@ require 'vbms/common'
 require 'vbms/client'
 require 'vbms/version'
 require 'vbms/requests'
-require 'vbms/soap-scum'
+require 'vbms/soap_scum'
 
 require 'vbms/responses/document'
 require 'vbms/responses/document_type'
