+++ lib/vbms/soap_scum.rb
